Add support for cumulative predictors
Adds support for positional arguments -- slices, maps.

Kong can append all remaining arguments into a slice or map if defined.
We can detect the presence of a cumulative positional argument if there exists
positional arguments and the last one is cumulative (either map or set).

If positional arguments is either a map or set, we may continue to use the
last positional predictor that is defined.

closes #16